% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_auth.R
\name{awsvault_profile}
\alias{awsvault_profile}
\title{AWS Vault Profile Switch}
\usage{
awsvault_profile(p)
}
\arguments{
\item{p}{Name of profile from \verb{~/.aws/config} to switch to.}
}
\description{
AWS Vault Profile Switch
}
\examples{
\dontrun{
awsvault_profile("upro")
}
}
