% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasharing.R
\name{datashare_um}
\alias{datashare_um}
\title{Datashare umccrise Results}
\usage{
datashare_um(sid, lid, token_ica)
}
\arguments{
\item{sid}{SubjectID.}

\item{lid}{LibraryID of WTS tumor.}

\item{token_ica}{ICA_ACCESS_TOKEN.}
}
\value{
Tibble with presigned URLs.
}
\description{
Datashare umccrise Results
}
