% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasharing.R
\name{datashare_um_s3}
\alias{datashare_um_s3}
\title{Datashare umccrise S3 Results}
\usage{
datashare_um_s3(libid, token)
}
\arguments{
\item{sid}{SubjectID.}

\item{lid}{LibraryID of WGS tumor.}
}
\value{
Tibble with presigned URLs.
}
\description{
Datashare umccrise S3 Results
}
\examples{
\dontrun{
libid <- "L2401591"
token <- orca_jwt() |> jwt_validate()
datashare_um_s3(libid, token)
}
}
