% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasharing.R
\name{datashare_wts_s3}
\alias{datashare_wts_s3}
\title{Datashare WTS S3 Results}
\usage{
datashare_wts_s3(libid, token)
}
\arguments{
\item{libid}{LibraryID of WTS tumor.}
}
\value{
Tibble with presigned URLs.
}
\description{
Datashare WTS S3 Results
}
\examples{
\dontrun{
libid <- "L2401585"
token <- orca_jwt() |> jwt_validate()
datashare_wts_s3(libid, token)
}
}
