% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{envvar_defined}
\alias{envvar_defined}
\title{Are EnvVars Defined?}
\usage{
envvar_defined(vars)
}
\arguments{
\item{vars}{Vector of env vars to check.}
}
\value{
Tibble with undefined env vars.
}
\description{
Are EnvVars Defined?
}
\examples{
envvar_defined("HOME")
}
