% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{glims_excel_read}
\alias{glims_excel_read}
\title{Read Google LIMS}
\usage{
glims_excel_read()
}
\value{
Tibble with all columns and rows from the Google LIMS spreadsheet.
}
\description{
Reads UMCCR's Google LIMS spreadsheet.
}
