% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{glims_portal_read}
\alias{glims_portal_read}
\title{Read Portal Limsrow Table}
\usage{
glims_portal_read(libids)
}
\arguments{
\item{libids}{Character vector of LibraryID values to query for.}
}
\value{
Tibble with all rows from the data portal limsrow table where
there are hits with the \code{library_id} column.
}
\description{
Reads rows from the data portal's limsrow table, given a set of \code{LibraryID}s
to query.
}
