% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_auth.R
\name{jwt_validate}
\alias{jwt_validate}
\title{JWT Validation}
\usage{
jwt_validate(token)
}
\arguments{
\item{token}{JWT string.}
}
\value{
The input token if valid, else errors out.
}
\description{
Validates JWT by parsing it and checking its structure and expiration date.
}
