% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_icav1.R
\name{meta_bcl_convert}
\alias{meta_bcl_convert}
\title{Metadata for bcl_convert workflow}
\usage{
meta_bcl_convert(pmeta, status = "Succeeded")
}
\arguments{
\item{pmeta}{Path to portal workflows metadata table, or tibble with already parsed data.}

\item{status}{Workflow status to keep (default: Succeeded).}
}
\value{
A tibble with metadata per workflow run.
}
\description{
Metadata for bcl_convert workflow
}
\examples{
pmeta <- "extdata/portaldb_workflow_top4.rds" |>
  system.file(package = "rportal") |>
  readr::read_rds()
(m <- meta_bcl_convert(pmeta))
\dontrun{
q1 <- glue("WHERE \"type_name\" = 'bcl_convert' ORDER BY \"start\" DESC LIMIT 4;")
pmeta_raw <- portaldb_query_workflow(q1)
m2 <- meta_bcl_convert(pmeta)
}
}
