% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_oncoanalyser_wgts_existing_both.R
\name{meta_oncoanalyser_wgts_existing_both}
\alias{meta_oncoanalyser_wgts_existing_both}
\title{Metadata for oncoanalyser_wgts_existing_both workflow}
\usage{
meta_oncoanalyser_wgts_existing_both(pmeta, status = "Succeeded")
}
\arguments{
\item{pmeta}{Path to portal workflows metadata table, or tibble with already parsed data.}

\item{status}{Workflow status to keep (default: Succeeded).}
}
\value{
A tibble with metadata per workflow run.
}
\description{
Metadata for oncoanalyser_wgts_existing_both workflow
}
\examples{
pmeta <- "extdata/portaldb_workflow_top4.rds" |>
  system.file(package = "rportal") |>
  readr::read_rds()
(m <- meta_oncoanalyser_wgts_existing_both(pmeta))
}
