% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_rnasum.R
\name{meta_rnasum}
\alias{meta_rnasum}
\title{Metadata for rnasum workflow}
\usage{
meta_rnasum(pmeta, status = "Succeeded")
}
\arguments{
\item{pmeta}{Path to portal workflows metadata table, or tibble with already parsed data.}

\item{status}{Workflow status to keep (default: Succeeded).}
}
\value{
A tibble with metadata per workflow run.
}
\description{
Metadata for rnasum workflow
}
\examples{
pmeta <- "extdata/portaldb_workflow_top4.rds" |>
  system.file(package = "rportal") |>
  readr::read_rds()
(m <- meta_rnasum(pmeta))
}
