% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_tso_ctdna_tumor_only.R
\name{meta_tso_ctdna_tumor_only}
\alias{meta_tso_ctdna_tumor_only}
\title{Metadata for tso_ctdna_tumor_only workflow}
\usage{
meta_tso_ctdna_tumor_only(pmeta, status = c("Succeeded"))
}
\arguments{
\item{pmeta}{Path to portal workflows metadata table, or tibble with already parsed data.}

\item{status}{Workflow status to keep (default: Succeeded).}
}
\value{
A tibble with metadata per workflow run.
}
\description{
Metadata for tso_ctdna_tumor_only workflow
}
\examples{
pmeta <- system.file("extdata/portal_meta_top4.csv", package = "rportal")
(m <- meta_tso_ctdna_tumor_only(pmeta))
}
