% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_umccrise.R
\name{meta_umccrise}
\alias{meta_umccrise}
\title{Metadata for umccrise workflow}
\usage{
meta_umccrise(pmeta, status = "Succeeded")
}
\arguments{
\item{pmeta}{Path to portal workflows metadata table, or tibble with already parsed data.}

\item{status}{Workflow status to keep (default: Succeeded).}
}
\value{
A tibble with metadata per workflow run.
}
\description{
Metadata for umccrise workflow
}
\examples{
pmeta <- system.file("extdata/portal_meta_top4.csv", package = "rportal")
(m <- meta_umccrise(pmeta))
}
