% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcabus.R
\name{orca_meta_libid2extsampleid}
\alias{orca_meta_libid2extsampleid}
\title{Grab externalSampleId from libraryId}
\usage{
orca_meta_libid2extsampleid(libid, token, stage = "prod")
}
\arguments{
\item{libid}{The library ID.}

\item{token}{JWT to use for authentication.}

\item{stage}{Environment where API is deployed (prod, stg or dev).}
}
\value{
The external sample ID.
}
\description{
Grab externalSampleId from libraryId
}
\examples{
token <- orca_jwt() |> jwt_validate()
libid <- "L2600301"
orca_meta_libid2extsampleid(libid, token)
}
