% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcabus.R
\name{orca_prid2wfpayload}
\alias{orca_prid2wfpayload}
\title{OrcaBus PortalRunId To Workflow Payload}
\usage{
orca_prid2wfpayload(prid, token, stage = "prod")
}
\arguments{
\item{prid}{portalRunId.}

\item{token}{JWT.}

\item{stage}{Environment where API is deployed (prod, stg or dev).}
}
\value{
List of workflow payload.
}
\description{
Given an OrcaBus portalRunId, provides the workflow payload.
}
\examples{
\dontrun{
prid <- "20241110c01a1c76" # cttso
prid <- "2024111638b77605" # sash
prid <- "20241116dfee1aef" # oa-wgts-dna
prid <- "2024111514abc96a" # wgs-tn
prid <- "20241115cbfdaeea" # wgs-qc-rna
prid <- "202411154e2c74f3" # wgs-qc-dna
prid <- "202411231acb8163" # wgs-qc-dna
prid <- "2024111507e8ca78" # bclconvert
prid <- "202411152feba98c" # bclconvert-interopqc
prid <- "2025060866284123" # wgs-tn
prid <- "20250630a115127b" # dragen-wgts-dna (germline only)
prid <- "20250624d45e528c" # dragen-wgts-dna (?)
token <- orca_jwt() |> jwt_validate()
p3 <- orca_prid2wfpayload(prid = prid, token = token)
}

}
