% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcabus.R
\name{orca_prid2wfrid}
\alias{orca_prid2wfrid}
\title{OrcaBus PortalRunId To WFR ID}
\usage{
orca_prid2wfrid(prid, token, stage = "prod")
}
\arguments{
\item{prid}{portalRunId.}

\item{token}{JWT to use for authentication.}

\item{stage}{Environment where API is deployed (prod, stg or dev).}
}
\value{
Workflow run ID.
}
\description{
Given an OrcaBus portalRunId, provides the workflow run ID.
}
\examples{
\dontrun{
token <- orca_jwt() |> jwt_validate()
prid <- "20241110c01a1c76"
wfrid <- orca_prid2wfrid(prid = prid, token = token)
}

}
