% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcabus.R
\name{orca_query_url}
\alias{orca_query_url}
\title{OrcaBus Query URL}
\usage{
orca_query_url(url, token = NULL)
}
\arguments{
\item{url}{URL to query.}

\item{token}{JWT to use for authentication.}
}
\value{
JSON response.
}
\description{
OrcaBus Query URL
}
