% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcabus.R
\name{orca_workflow_list}
\alias{orca_workflow_list}
\title{OrcaBus List Workflow Runs}
\usage{
orca_workflow_list(wf_name = NULL, token, page_size = 10, stage = "prod")
}
\arguments{
\item{wf_name}{Name of workflow. If NULL, returns all.}

\item{token}{JWT.}

\item{page_size}{Maximum number of rows to return.}

\item{stage}{Environment where API is deployed (prod, stg or dev).}
}
\value{
Tibble with results.
}
\description{
OrcaBus List Workflow Runs
}
\examples{
\dontrun{
token <- orca_jwt() |> jwt_validate()
wf_name <- NULL
wf_name <- "cttsov2"
orca_workflow_list(wf_name = wf_name, token = token)
}
}
