% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcabus.R
\name{orca_workflow_payload}
\alias{orca_workflow_payload}
\title{OrcaBus cttsov2 PortalRunId To Output Directory}
\usage{
orca_workflow_payload(prid, token)
}
\arguments{
\item{prid}{PortalRunId.}

\item{token}{JWT.}
}
\value{
List with the sampleId \code{prefix} and the \code{path} containing the Results
and Logs_Intermediates folders.
}
\description{
Given a portalRunId from an OrcaBus cttso run, generate the sampleId and
root path to the output directory with the results.
}
\examples{
\dontrun{
token <- orca_jwt() |> jwt_validate()
# e.g. for a cttsov2 workflow
orca_workflow_payload("20240930fc23056d", token = token)

sampleid <- resp[["data"]][["inputs"]][["sampleId"]]
path <- resp[["data"]][["outputs"]][["resultsDir"]] |> dirname()
list(
  path = path,
  prefix = sampleid
)
}

}
