% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcabus_pld.R
\name{pld_bclconvert}
\alias{pld_bclconvert}
\title{Payload Tidy bclconvert}
\usage{
pld_bclconvert(pld)
}
\arguments{
\item{pld}{List with bclconvert workflow parameters.}
}
\value{
A tidy tibble.
}
\description{
Payload Tidy bclconvert
}
