% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcabus_pld.R
\name{pld_bclconvertinteropqc}
\alias{pld_bclconvertinteropqc}
\title{Payload Tidy bclconvertinteropqc}
\usage{
pld_bclconvertinteropqc(pld)
}
\arguments{
\item{pld}{List with bclconvertinteropqc workflow parameters.}
}
\value{
A tidy tibble.
}
\description{
Payload Tidy bclconvertinteropqc
}
