% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcabus_pld.R
\name{pld_bsshfastqcopy}
\alias{pld_bsshfastqcopy}
\title{Payload Tidy bsshfastqcopy}
\usage{
pld_bsshfastqcopy(pld)
}
\arguments{
\item{pld}{List with bsshfastqcopy workflow parameters.}
}
\value{
A tidy tibble.
}
\description{
Payload Tidy bsshfastqcopy
}
