% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcabus_pld.R
\name{pld_dragentso500ctdna}
\alias{pld_dragentso500ctdna}
\title{Payload Tidy dragen-tso500-ctdna - *Note there is an issue with "IndividualId" returned, does not match lims}
\usage{
pld_dragentso500ctdna(pld)
}
\arguments{
\item{pld}{List with dragen-tso500-ctdna (previously cttsov2) workflow parameters.}
}
\value{
A tidy tibble.
}
\description{
Payload Tidy dragen-tso500-ctdna - *Note there is an issue with "IndividualId" returned, does not match lims
}
