% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcabus_pld.R
\name{pld_dragenwgtsdna}
\alias{pld_dragenwgtsdna}
\title{Payload Tidy dragenwgtsdna}
\usage{
pld_dragenwgtsdna(pld)
}
\arguments{
\item{pld}{List with dragenwgtsdna workflow parameters.}
}
\value{
A tidy tibble.
}
\description{
Payload Tidy dragenwgtsdna
}
