% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcabus_pld.R
\name{pld_oawgtsdnarna}
\alias{pld_oawgtsdnarna}
\title{Payload Tidy oncoanalyser-wgts-dna-rna}
\usage{
pld_oawgtsdnarna(pld)
}
\arguments{
\item{pld}{List with oncoanalyserwgtsdnarna workflow parameters.}
}
\value{
A tidy tibble.
}
\description{
Payload Tidy oncoanalyser-wgts-dna-rna
}
