% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcabus_pld.R
\name{pld_oawgtsrna}
\alias{pld_oawgtsrna}
\title{Payload Tidy oncoanalyser-wgts-rna}
\usage{
pld_oawgtsrna(pld)
}
\arguments{
\item{pld}{List with oncoanalyserwgtsrna workflow parameters.}
}
\value{
A tidy tibble.
}
\description{
Payload Tidy oncoanalyser-wgts-rna
}
