% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcabus_pld.R
\name{pld_oracompression}
\alias{pld_oracompression}
\title{Payload Tidy orca-compression}
\usage{
pld_oracompression(pld)
}
\arguments{
\item{pld}{List with orcacompression workflow parameters.}
}
\value{
A tidy tibble.
}
\description{
Payload Tidy orca-compression
}
