% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcabus_pld.R
\name{pld_pieriandx}
\alias{pld_pieriandx}
\title{Payload Tidy pieriandx}
\usage{
pld_pieriandx(pld)
}
\arguments{
\item{pld}{List with pieriandx workflow parameters.}
}
\value{
A tidy tibble.
}
\description{
Payload Tidy pieriandx
}
