% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcabus_pld.R
\name{pld_rnasum}
\alias{pld_rnasum}
\title{Payload Tidy rnasum}
\usage{
pld_rnasum(pld)
}
\arguments{
\item{pld}{List with rnasum workflow parameters.}
}
\value{
A tidy tibble.
}
\description{
Payload Tidy rnasum
}
