% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcabus_pld.R
\name{pld_tumornormal}
\alias{pld_tumornormal}
\title{Payload Tidy tumor_normal}
\usage{
pld_tumornormal(pld)
}
\arguments{
\item{pld}{List with tumornormal workflow parameters.}
}
\value{
A tidy tibble.
}
\description{
Payload Tidy tumor_normal
}
