% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_wgs_tumor_normal.R
\name{pld_wgstn}
\alias{pld_wgstn}
\title{Payload Tidy wgstn}
\usage{
pld_wgstn(pld)
}
\arguments{
\item{pld}{List with wgs-tn workflow parameters.}
}
\value{
A tidy tibble.
}
\description{
Payload Tidy wgstn
}
