% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_wgs_alignment_qc.R
\name{pld_wgtsqc}
\alias{pld_wgtsqc}
\title{Payload Tidy wgtsqc}
\usage{
pld_wgtsqc(pld)
}
\arguments{
\item{pld}{List with wgtsqc workflow parameters.}
}
\value{
A tidy tibble.
}
\description{
Payload Tidy wgtsqc
}
