% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_wts_tumor_only.R
\name{pld_wts}
\alias{pld_wts}
\title{Payload for WTS workflow}
\usage{
pld_wts(pld)
}
\arguments{
\item{pld}{List with WTS workflow parameters.}
}
\value{
A tidy tibble.
}
\description{
Payload for WTS workflow
}
