% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{portal_meta_read}
\alias{portal_meta_read}
\title{Read ICA Workflows Metadata via Portal API}
\usage{
portal_meta_read(pmeta = NULL, rows = 100, params = "", account = "prod")
}
\arguments{
\item{pmeta}{Path to downloaded portal metadata file, or already parsed metadata tibble.}

\item{rows}{Number of rows to return.}

\item{params}{String containing additional params to pass to the \verb{/workflows}
endpoint, e.g. \code{'&type_name=bclconvert'}.}

\item{account}{UMCCR portal account (one of "prod", "dev", "stg").}
}
\value{
A tibble of the results from the given query.
}
\description{
Reads ICA Workflows Metadata via Portal API using awscurl. See
https://github.com/okigan/awscurl for required \code{AWS_} environment variables.
}
\examples{
\dontrun{
portal_meta_read(params = "&type_name=rnasum", rows = 4)
}
}
