% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{portal_meta_read_athena}
\alias{portal_meta_read_athena}
\title{Read ICA Workflows Metadata via Athena}
\usage{
portal_meta_read_athena(wfrids = NULL)
}
\arguments{
\item{wfrids}{Character vector of wfr IDs to query.}
}
\value{
Tibble with a row per wfr ID.
}
\description{
Reads the ICA Workflows Metadata for the given workflow run IDs.
}
\examples{
\dontrun{
wfrids <- c("wfr.1e764ca00e7a43a69e2424f250a34868")
portal_meta_read_athena(wfrids)
}
}
