% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{portaldb_connect}
\alias{portaldb_connect}
\title{PortalDB Connect}
\usage{
portaldb_connect(workgroup = "orcahouse")
}
\arguments{
\item{workgroup}{UMCCR Athena workgroup for portal database.}
}
\value{
Connection to DBMS
}
\description{
Establish the connection parameters for querying the Athena UMCCR databases
from the \code{orcahouse} or \code{data_portal} workgroup.
}
\examples{
\dontrun{
con <- portaldb_connect(workgroup = "orcahouse")
RAthena::list_work_groups(con) |> str()
}
}
