% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{portaldb_query}
\alias{portaldb_query}
\title{PortalDB Query}
\usage{
portaldb_query(query = NULL, workgroup = "orcahouse")
}
\arguments{
\item{query}{SQL query string.}

\item{workgroup}{UMCCR Athena workgroup for portal database.}
}
\value{
Tibble with results from the provided query.
}
\description{
Query the \code{mart} UMCCR database from the \code{orcahouse} workgroup.
To access a specific table from the database, use
\verb{"orcavault"."mart"."table"} (see examples).
}
\examples{
\dontrun{
q1 <- 'SELECT * FROM "orcavault"."mart"."lims" LIMIT 2;'
portaldb_query(query = q1)
# example of old query from data_portal database
q2 <- 'SELECT * FROM "data_portal"."data_portal"."data_portal_limsrow" LIMIT 2;'
portaldb_query(query = q2, workgroup = "data_portal")
# example of finding exact matches
prids <- c("20240327d595afa9", "202403274bf3ad80", "202504181b2efa22")
prids_quote <- paste(shQuote(prids), collapse = ", ")
wf_table <- '"orcavault"."mart"."workflow"'
q3 <- glue('SELECT * FROM {wf_table} WHERE "portal_run_id" IN ({prids_quote});')
portaldb_query(q3)
}
}
