% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{portaldb_query_accreditationlims}
\alias{portaldb_query_accreditationlims}
\title{PortalDB Query Mart accreditation_lims Table}
\usage{
portaldb_query_accreditationlims(query = NULL)
}
\arguments{
\item{query}{SQL query string.}
}
\value{
Tibble with results from query.
}
\description{
Queries the \code{accreditation_lims} table with the given query.
Note this is simply a convenience function that prepends
\verb{SELECT * FROM "orcavault"."mart"."accreditation_lims" } to any
\code{query} you provide.
See examples.
}
\examples{
\dontrun{
libids <- shQuote(paste(c("L2501116", "L2501087", "L2501069"), collapse = "|"))
query <- glue("WHERE REGEXP_LIKE(\"library_id\", {libids});")
portaldb_query_accreditationlims(query)
query_ptc <- glue("WHERE \"label\" = 'ptc';")
portaldb_query_accreditationlims(query_ptc)
}
}
