% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{portaldb_query_bam}
\alias{portaldb_query_bam}
\title{PortalDB Query bam Table}
\usage{
portaldb_query_bam(query = NULL)
}
\arguments{
\item{query}{SQL query string.}
}
\value{
Tibble with results from query.
}
\description{
Queries the \code{bam} table with the given query.
Note this is simply a convenience function that prepends
\verb{SELECT * FROM "orcavault"."mart"."bam" } to any
\code{query} you provide.
See examples.
}
\examples{
\dontrun{
libids <- shQuote(paste(c("L2400340", "L2400256", "L2500469"), collapse = "|"))
query <- glue("WHERE REGEXP_LIKE(\"library_id\", {libids});")
portaldb_query_bam(query)
}
}
