% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{portaldb_query_fastqhistory}
\alias{portaldb_query_fastqhistory}
\title{PortalDB Query fastq_history Table}
\usage{
portaldb_query_fastqhistory(query = NULL)
}
\arguments{
\item{query}{SQL query string.}
}
\value{
Tibble with results from query.
}
\description{
Queries the \code{fastq_history} table with the given query.
Note this is simply a convenience function that prepends
\verb{SELECT * FROM "orcavault"."mart"."fastq_history" } to any
\code{query} you provide.
See examples.
}
\examples{
\dontrun{
libids <- shQuote(paste(c("L2100192", "L2100191", "L2500469"), collapse = "|"))
query <- glue("WHERE REGEXP_LIKE(\"library_id\", {libids});")
res <- portaldb_query_fastqhistory(query)
}
}
