% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{portaldb_query_fastqlistrow}
\alias{portaldb_query_fastqlistrow}
\title{PortalDB Query Fastqlistrow Table}
\usage{
portaldb_query_fastqlistrow(query = NULL)
}
\arguments{
\item{query}{SQL query string.}
}
\value{
Tibble with results from query.
}
\description{
Queries the \code{data_portal_fastqlistrow} table with the given query.
Note this is simply a convenience function that prepends
\verb{SELECT * FROM "data_portal"."data_portal"."data_portal_fastqlistrow" } to any
\code{query} you provide.
See examples.
}
\examples{
\dontrun{
libids <- shQuote(paste(c("L2100192", "L2100191"), collapse = "|"))
query <- glue("WHERE REGEXP_LIKE(\"rglb\", {libids});")
res <- portaldb_query_fastqlistrow(query)
}
}
