% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{portaldb_query_lims}
\alias{portaldb_query_lims}
\title{PortalDB Query Mart lims Table}
\usage{
portaldb_query_lims(query = NULL)
}
\arguments{
\item{query}{SQL query string.}
}
\value{
Tibble with results from query.
}
\description{
Queries the \code{lims} table with the given query.
Note this is simply a convenience function that prepends
\verb{SELECT * FROM "orcavault"."mart"."lims" } to any
\code{query} you provide.
See examples.
}
\examples{
\dontrun{
libids <- shQuote(paste(c("L2400340", "L2400256", "L2500469"), collapse = "|"))
query1 <- glue("WHERE REGEXP_LIKE(\"library_id\", {libids});")
portaldb_query_lims(query1)
sbjids <- paste(c("SBJ04470", "SBJ04487", "SBJ04488"), collapse = "|")
query2 <- glue(
  "WHERE REGEXP_LIKE(\"internal_subject_id\", '{sbjids}') AND \"type\" = 'WGS' ",
  "AND \"phenotype\" = 'tumor' ORDER BY \"internal_subject_id\" DESC;"
)
d <- portaldb_query_lims(query2)
# get tumor libids for each sbjid
d |> dplyr::select(internal_subject_id, library_id)
}
}
