% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{portaldb_query_limsrow}
\alias{portaldb_query_limsrow}
\title{PortalDB Query Limsrow Table}
\usage{
portaldb_query_limsrow(query = NULL)
}
\arguments{
\item{query}{SQL query string.}
}
\value{
Tibble with results from query.
}
\description{
Queries the \code{data_portal_limsrow} table with the given query.
Note this is simply a convenience function that prepends
\verb{SELECT * FROM "data_portal"."data_portal"."data_portal_limsrow" } to any
\code{query} you provide.
See examples.
}
\examples{
\dontrun{
libids <- shQuote(paste(c("L2400340", "L2400256"), collapse = "|"))
query1 <- glue("WHERE REGEXP_LIKE(\"library_id\", {libids});")
portaldb_query_limsrow(query1)
sbjids <- paste(c("SBJ04470", "SBJ04487", "SBJ04488"), collapse = "|")
query2 <- glue(
  "WHERE REGEXP_LIKE(\"subject_id\", '{sbjids}') AND \"type\" = 'WGS' ",
  "AND \"phenotype\" = 'tumor' ORDER BY \"subject_id\" DESC;"
)
d <- portaldb_query_limsrow(query2)
# get tumor libids for each sbjid
d |> dplyr::select(subject_id, library_id)
}
}
