% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{portaldb_query_workflow}
\alias{portaldb_query_workflow}
\title{PortalDB Query Workflow Table}
\usage{
portaldb_query_workflow(query = NULL)
}
\arguments{
\item{query}{SQL query string.}
}
\value{
Tibble with results from query.
}
\description{
Queries the \code{data_portal_workflow} table with the given query.
Note this is simply a convenience function that prepends
\verb{SELECT * FROM "data_portal"."data_portal"."data_portal_workflow" } to any
\code{query} you provide.
See examples.
}
\examples{
\dontrun{
prid <- shQuote("202403274bf3ad80")
query <- glue("WHERE \"portal_run_id\" IN ({prid});")
portaldb_query_workflow(query)
}
}
