% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rportal-package.R
\docType{package}
\name{rportal-package}
\alias{rportal}
\alias{rportal-package}
\title{rportal: UMCCR Portal R Functions}
\description{
Contains R functionality for interacting with the UMCCR Data Portal.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/umccr/rportal}
  \item \url{https://umccr.github.io/rportal/}
  \item Report bugs at \url{https://github.com/umccr/rportal/issues}
}

}
\author{
\strong{Maintainer}: Peter Diakumis \email{peterdiakumis@gmail.com} (\href{https://orcid.org/0000-0002-7502-7545}{ORCID})

}
\keyword{internal}
