% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_bcl_convert.R
\name{meta_bcl_convert}
\alias{meta_bcl_convert}
\title{Metadata for bcl_convert workflow}
\usage{
meta_bcl_convert(pmeta, status = "Succeeded")
}
\arguments{
\item{pmeta}{Path to portal workflows metadata table, or tibble with already parsed data.}

\item{status}{Workflow status to keep (default: Succeeded).}
}
\value{
A tibble with metadata per workflow run.
}
\description{
Metadata for bcl_convert workflow
}
\examples{
pmeta <- system.file("extdata/portal_meta_top4.csv", package = "rportal")
(m <- meta_bcl_convert(pmeta))
}
