% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_icav1.R
\name{meta_umccrise}
\alias{meta_umccrise}
\title{Metadata for umccrise workflow}
\usage{
meta_umccrise(pmeta, status = "Succeeded")
}
\arguments{
\item{pmeta}{tibble with raw PortalDB workflow metadata.}

\item{status}{Workflow status to keep (default: Succeeded).}
}
\value{
A tibble with tidy metadata per workflow run.
}
\description{
Metadata for umccrise workflow
}
\examples{
pmeta <- "extdata/portaldb_workflow_top4.rds" |>
  system.file(package = "rportal") |>
  readr::read_rds()
(m <- meta_umccrise(pmeta))
}
