% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcabus.R
\name{orca_libid2workflows}
\alias{orca_libid2workflows}
\title{OrcaBus Get Workflows From LibraryId}
\usage{
orca_libid2workflows(
  libid,
  token,
  wf_name = NULL,
  page_size = 10,
  stage = "prod"
)
}
\arguments{
\item{libid}{libraryId.}

\item{token}{JWT.}

\item{wf_name}{Name of workflow to query.}

\item{page_size}{Maximum number of rows to return.}

\item{stage}{Environment where API is deployed (prod, stg or dev).}
}
\value{
Tibble with results.
}
\description{
Given a libraryId gets workflow details.
}
\examples{
\dontrun{
token <- orca_jwt() |> jwt_validate()
libid <- "L2401591"
wf_name <- NULL
d <- orca_libid2workflows(libid = libid, token = token, wf_name = wf_name)
}
}
