% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcabus.R
\name{orca_prid2wfpayload}
\alias{orca_prid2wfpayload}
\title{OrcaBus PortalRunId To Workflow Payload}
\usage{
orca_prid2wfpayload(prid, token, stage = "prod")
}
\arguments{
\item{prid}{portalRunId.}

\item{token}{JWT.}

\item{stage}{Environment where API is deployed (prod, stg or dev).}
}
\value{
List of workflow payload.
}
\description{
Given an OrcaBus portalRunId, provides the workflow payload.
}
\examples{
\dontrun{
token <- orca_jwt() |> jwt_validate()
# e.g. for a cttsov2 workflow
prid <- "20240930a40b3974"
p <- orca_prid2wfpayload(prid = prid, token = token)

sampleid <- resp[["data"]][["inputs"]][["sampleId"]]
path <- resp[["data"]][["outputs"]][["resultsDir"]] |> dirname()
list(
  path = path,
  prefix = sampleid
)
}

}
