% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcabus.R
\name{orca_wfrid2payload}
\alias{orca_wfrid2payload}
\title{OrcaBus WFR ID To Payload}
\usage{
orca_wfrid2payload(wfrid, token, stage = "prod")
}
\arguments{
\item{wfrid}{Workflow run ID.}

\item{token}{JWT to use for authentication.}

\item{stage}{Environment where API is deployed (prod, stg or dev).}
}
\value{
List of workflow payload.
}
\description{
Given a workflow run ID, provides the payload.
}
\examples{
\dontrun{
token <- orca_jwt() |> jwt_validate()
wfrid <- "wfr.01JCARAVTXKG5581SRA1HKBTD3"
wfrid <- "wfr.01JCA5DZFD0T4MFQX0HHEEFBCH" # wts
wfrid <- "wfr.01JBX361HKV0V9WS96RAFG135T" # cttsov2
p <- orca_wfrid2payload(wfrid = wfrid, token = token)
}

}
