% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcabus_pld.R
\name{pld_cttsov2}
\alias{pld_cttsov2}
\title{Payload Tidy tso}
\usage{
pld_cttsov2(pld)
}
\arguments{
\item{pld}{List with cttsov2 workflow parameters.}
}
\value{
A tidy tibble.
}
\description{
Payload Tidy tso
}
