% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcabus_pld.R
\name{pld_oawgtsdna}
\alias{pld_oawgtsdna}
\title{Payload Tidy oncoanalyser-wgts-dna}
\usage{
pld_oawgtsdna(pld)
}
\arguments{
\item{pld}{List with oncoanalyserwgtsdna workflow parameters.}
}
\value{
A tidy tibble.
}
\description{
Payload Tidy oncoanalyser-wgts-dna
}
