% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_umccrise.R
\name{pld_umccrise}
\alias{pld_umccrise}
\title{Payload for umccrise workflow}
\usage{
pld_umccrise(pld)
}
\arguments{
\item{pld}{List with umccrise workflow parameters.}
}
\value{
A tidy tibble.
}
\description{
Payload for umccrise workflow
}
