% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{portaldb_query}
\alias{portaldb_query}
\title{PortalDB Query}
\usage{
portaldb_query(query = NULL)
}
\arguments{
\item{query}{SQL query string.}
}
\value{
Tibble with results from the provided query.
}
\description{
Query the \code{data_portal} UMCCR database.
To access a specific table from the database, use
\verb{"data_portal"."data_portal"."table"} (see examples).
}
\examples{
\dontrun{
q1 <- 'SELECT * FROM "data_portal"."data_portal"."data_portal_limsrow" LIMIT 2;'
portaldb_query(q1)
# example of finding exact matches
prids <- c("20240327d595afa9", "202403274bf3ad80")
prids_quote <- paste(shQuote(prids), collapse = ", ")
wf_table <- '"data_portal"."data_portal"."data_portal_workflow"'
q2 <- glue('SELECT * FROM {wf_table} WHERE "portal_run_id" IN ({prids_quote});')
portaldb_query(q2)
}
}
