% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{portaldb_query}
\alias{portaldb_query}
\title{PortalDB Query}
\usage{
portaldb_query(query = NULL, workgroup = "orcahouse")
}
\arguments{
\item{query}{SQL query string.}
}
\value{
Tibble with results from the provided query.
}
\description{
Submit an Athena query.
}
\examples{
\dontrun{
mart_db <- '"orcavault"."mart"'
q1 <- glue('SELECT * FROM {mart_db}."lims" LIMIT 2;')
portaldb_query(query = q1)
# find exact matches
prids <- c("20240327d595afa9", "202403274bf3ad80", "202504181b2efa22")
prids_quote <- paste(shQuote(prids), collapse = ", ")
q2 <- glue('SELECT * FROM {mart_db}."workflow" WHERE "portal_run_id" IN ({prids_quote});')
portaldb_query(q2)
}
}
