% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{portaldb_query_workflow}
\alias{portaldb_query_workflow}
\title{PortalDB Query Mart workflow Table}
\usage{
portaldb_query_workflow(query = NULL)
}
\arguments{
\item{query}{SQL query string.}
}
\value{
Tibble with results from query.
}
\description{
Queries the \code{workflow} table with the given query.
Note this is simply a convenience function that prepends
\verb{SELECT * FROM "orcavault"."mart"."workflow" } to any
\code{query} you provide.
See examples.
}
\examples{
\dontrun{
prid <- shQuote("2025060866284123")
query <- glue("WHERE \"portal_run_id\" IN ({prid});")
query <- glue("WHERE \"workflow_name\" = 'dragen-wgts-dna';")
query <- ";"
wfs <- portaldb_query_workflow(query)
}
}
